backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.53
marshmallow==3.20.1
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.3.1
pytest-asyncio==0.21.1
httpx>=0.27.0